+++ tests/projections/data/equal-area-test-data.ts
