+++ tests/projections/equal-area.test.ts
@@ -2,7 +2,7 @@ import {glMatrix, vec3} from 'gl-matrix';
 glMatrix.setMatrixArrayType(Float64Array as any);
 import {describe, it, expect} from 'vitest';
 import {EqualAreaProjection} from '../../modules/projections/equal-area';
-import TEST_DATA from './fixtures/polyhedral.json';
+import TEST_DATA from './fixtures/equal-area.json';
 import {Cartesian} from '../../modules/core/coordinate-systems';
 
 // Extract static data from test data
@@ -18,12 +18,12 @@ const MAX_ARC_LENGTH_MM = AUTHALIC_RADIUS * MAX_ANGLE * 1e9;
 const DESIRED_MM_PRECISION = 0.01;
 
 describe('EqualAreaProjection forward', () => {
-  const polyhedral = new EqualAreaProjection();
+  const equalArea = new EqualAreaProjection();
   let largestError = 0;
 
   it('forward projections', () => {
     TEST_DATA.forward.forEach((testCase, index) => {
-      const result = polyhedral.forward(
+      const result = equalArea.forward(
         testCase.input as any,
         TEST_SPHERICAL_TRIANGLE as any,
         TEST_FACE_TRIANGLE as any
@@ -35,8 +35,8 @@ describe('EqualAreaProjection forward', () => {
   it('round trip forward projections', () => {
     TEST_DATA.forward.forEach((testCase, index) => {
       const spherical = testCase.input as any;
-      const polar = polyhedral.forward(spherical, TEST_SPHERICAL_TRIANGLE as any, TEST_FACE_TRIANGLE as any);
-      const result = polyhedral.inverse(polar, TEST_FACE_TRIANGLE as any, TEST_SPHERICAL_TRIANGLE as any);
+      const polar = equalArea.forward(spherical, TEST_SPHERICAL_TRIANGLE as any, TEST_FACE_TRIANGLE as any);
+      const result = equalArea.inverse(polar, TEST_FACE_TRIANGLE as any, TEST_SPHERICAL_TRIANGLE as any);
       largestError = Math.max(largestError, vec3.distance(result, spherical));
       expect(result).toBeCloseToArray(spherical);
     });
@@ -48,11 +48,11 @@ describe('EqualAreaProjection forward', () => {
 });
 
 describe('EqualAreaProjection inverse', () => {
-  const polyhedral = new EqualAreaProjection();
+  const equalArea = new EqualAreaProjection();
 
   it('inverse projections', () => {
     TEST_DATA.inverse.forEach((testCase, index) => {
-      const result = polyhedral.inverse(
+      const result = equalArea.inverse(
         testCase.input as any,
         TEST_FACE_TRIANGLE as any,
         TEST_SPHERICAL_TRIANGLE as any
@@ -64,8 +64,8 @@ describe('EqualAreaProjection inverse', () => {
   it('round trip inverse projections', () => {
     TEST_DATA.inverse.forEach((testCase, index) => {
       const facePoint = testCase.input as any;
-      const spherical = polyhedral.inverse(facePoint, TEST_FACE_TRIANGLE as any, TEST_SPHERICAL_TRIANGLE as any);
-      const result = polyhedral.forward(spherical, TEST_SPHERICAL_TRIANGLE as any, TEST_FACE_TRIANGLE as any);
+      const spherical = equalArea.inverse(facePoint, TEST_FACE_TRIANGLE as any, TEST_SPHERICAL_TRIANGLE as any);
+      const result = equalArea.forward(spherical, TEST_SPHERICAL_TRIANGLE as any, TEST_FACE_TRIANGLE as any);
       expect(result).toBeCloseToArray(facePoint);
     });
   });
