+++ tests/projections/fixtures/equal-area.json
